requirements.txt
opencv-python
pytesseract
Pillow
numpy
scikit-image
fastapi
uvicorn[standard]
python-multipart
aiofiles